fixed: change checkoutRequest's quantity field from uint to float64 to support fractions on products with kg unit

+++ models/checkoutResponseRequestmodels.go
@@ -1,9 +1,9 @@
 package models
 
 type CartlineInput struct {
-	ProductID uint   `json:"productID"`
-	Unit      string `json:"unit"`
-	Quantity  uint   `json:"quantity"`
+	ProductID uint    `json:"productID"`
+	Unit      string  `json:"unit"`
+	Quantity  float64 `json:"quantity"`
 }
 
 type CheckOutrequest struct {
